refactor: split iron proxy config model
