+++ services/api-rs/crates/centaur-iron-proxy/src/model/mod.rs
@@ -1,0 +1,15 @@
+mod broker;
+mod postgres;
+mod proxy;
+mod transform;
+mod values;
+
+pub use broker::BrokerCredential;
+pub use postgres::{PgDsnEnv, PostgresClient, PostgresListener, PostgresUpstream, SandboxEnv};
+pub use proxy::ProxyFragment;
+pub use transform::{Secret, SecretReplace, Transform, TransformConfig};
+
+pub(crate) use proxy::ProxyConfig;
+pub(crate) use values::{
+    listen_port, non_empty, resolve_placeholder_source_values, value_field_str,
+};
